+++ additionalInstallationScripts/installWazuh.sh
@@ -450,3 +450,41 @@ main() {
 }
 
 main
+
+#######################################
+#wait until elasticsearch comes up before continuing
+ES_URL=${ES_URL:-'http://localhost:9200'}
+ES_USER=${ES_USER:-kibana}
+ES_PASSWORD=${ES_PASSWORD:-changeme}
+until curl -u ${ES_USER}:${ES_PASSWORD} -XGET "${ES_URL}"; do
+  service elasticsearch restart
+  sleep 5
+done
+>&2 echo "Elastic is up - executing commands"
+#######################################
+#set default kibana index to wazuh alerts
+K_URL='localhost:5601/api/kibana/settings/defaultIndex'
+K_USER='elastic'
+K_PASSWORD='changeme'
+# wait until kibana service is avilable.
+until curl -u ${K_USER}:${K_PASSWORD} -XGET "localhost:5601"; do
+  sleep 5
+done
+# wait until kibana service is ready
+until [ "$(curl "http://${K_USER}:${K_PASSWORD}@localhost:5601")" != "Kibana server is not ready yet" ]; do
+  sleep 5
+done
+echo "Kibana is up"
+curl -X POST -H "Content-Type: application/json" -H "kbn-xsrf: true" -d '{"value":"wazuh-alerts-3.x-*"}' "http://${K_USER}:${K_PASSWORD}@${K_URL}"
+############################################################################################################################################################
+# Import AWS Detonation lab dashboards
+K_URL='localhost:5601/api/kibana/dashboards/import'
+K_USER='elastic'
+K_PASSWORD='changeme'
+curl -o Kibana-Visualizations.json https://raw.githubusercontent.com/sonofagl1tch/AWSDetonationLab/master/KibanaAdditionalConfigs/Kibana-Visualizations.json
+curl -o Kibana-Dashboard.json https://raw.githubusercontent.com/sonofagl1tch/AWSDetonationLab/master/KibanaAdditionalConfigs/Kibana-Dashboard.json
+curl -X POST -H "Content-Type: application/json" -H "kbn-xsrf: true" "http://${K_USER}:${K_PASSWORD}@${K_URL}" -d @Kibana-Dashboard.json
+curl -X POST -H "Content-Type: application/json" -H "kbn-xsrf: true" "http://${K_USER}:${K_PASSWORD}@${K_URL}" -d @Kibana-Visualizations.json
+#######################################
+# next steps is to configure wazuh
+## https://documentation.wazuh.com/current/installation-guide/installing-elastic-stack/connect_wazuh_app.html