+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## v2.1
+### Added
+* Automate setting `wazuh-alerts-3.x-*` as Kibana's default index pattern ([#64](https://github.com/sonofagl1tch/AWSDetonationLab/pull/64/)).
+* Automate import of custom dashboards and visualizations into Kibana ([#68](https://github.com/sonofagl1tch/AWSDetonationLab/pull/68/)).
+
+### Fixed
+* Retrieve AMI IDs dynamically instead of having hardcoded values ([#66](https://github.com/sonofagl1tch/AWSDetonationLab/pull/66)).
+
 ## v2.0
 ### Added
 * Added `apache` user to `wheel` group in Linux vulnerable server ([#20](https://github.com/sonofagl1tch/AWSDetonationLab/pull/20)).